Update test setup for browsers
Update local test setup to use Node for scripts and chrome for the app. Then add a browserstack setup for travis.
