+++ .travis.yml
@@ -13,7 +13,7 @@ env:
 install:
 - travis_retry npm install
 script:
-- npm run lint && npm test && npm run uploadCoverage
+- npm run lint && npm run test:ci && npm run uploadCoverage
 notifications:
   slack:
     secure: SUW1XF8Nggc5m+Xw3U7qaMFM9Tfn0NmvZePI84K9INDlLx+YxVZouteKN0XW9WsgJbI0ZiJE0CthJRS/twcjW4hO/LaizXYbqQaC+3pEUGajgYL1rbpqr0BfKmPKplvR/LbbTMo/GfUOOmU5F5ajWh1HWadHp1Dj/GSPzms97FLq0Xs0bWe/saX17v69V+36X7jyIrWPdQ3nVH284bbLMKnlNQ7t0LuPD748m1neLGNdQTC2R7P2eFoRS5hR61qY4m2my96wH2PndVbt1HuTW2S61wAX3DCQ+5a21bTnv+VCawvhlEZq3/zgr9w0M7WAeFqbdA7gwvaAV3N80JAdvwX60Wfw4rfwADTwj/KAAgF3dbc4iYTwUunCjZKydgUn1WWIG2sLBb0u2GfXAWINOWeDm2tZPPFx9nB/8Fjxnasp85F9TXxhyMXkybKaV0hf57rjaGTDXkl3PxzFTOnpakZU/+Gv68UhO+VwDjuQakBuCg43xV9jrVHWZ941LP13bBhNh2u3UKJafrBQLf6SchbQyx8/iWwLgHNuHpskKjoHIo/bOTqFaHHo6o4UHnT5KeLdXLvyxM5pg7UyYHrT1T4kUzbVQtWGCba5EI+H3A11RLwKjKir4utdMsZgkJ8+ORNW499OOlk8Hfs2tGs/EbH9yxGTWzLEJWiu5LWJt5Y=
