+++ npm-debug.log
@@ -1,0 +1,49 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/Users/kanefreeman/.nvm/versions/node/v6.14.4/bin/node',
+1 verbose cli   '/Users/kanefreeman/.nvm/versions/node/v6.14.4/bin/npm',
+1 verbose cli   'run',
+1 verbose cli   'test:ci' ]
+2 info using npm@3.10.10
+3 info using node@v6.14.4
+4 verbose run-script [ 'pretest:ci', 'test:ci', 'posttest:ci' ]
+5 info lifecycle dojo.io@1.0.0~pretest:ci: dojo.io@1.0.0
+6 silly lifecycle dojo.io@1.0.0~pretest:ci: no script for pretest:ci, continuing
+7 info lifecycle dojo.io@1.0.0~test:ci: dojo.io@1.0.0
+8 verbose lifecycle dojo.io@1.0.0~test:ci: unsafe-perm in lifecycle true
+9 verbose lifecycle dojo.io@1.0.0~test:ci: PATH: /Users/kanefreeman/.nvm/versions/node/v6.14.4/lib/node_modules/npm/bin/node-gyp-bin:/Users/kanefreeman/Documents/GitHub/site/node_modules/.bin:/Users/kanefreeman/.nvm/versions/node/v6.14.4/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin
+10 verbose lifecycle dojo.io@1.0.0~test:ci: CWD: /Users/kanefreeman/Documents/GitHub/site
+11 silly lifecycle dojo.io@1.0.0~test:ci: Args: [ '-c',
+11 silly lifecycle   'run-s test:build test:build:static intern:ci test:clean' ]
+12 silly lifecycle dojo.io@1.0.0~test:ci: Returned: code: 1  signal: null
+13 info lifecycle dojo.io@1.0.0~test:ci: Failed to exec test:ci script
+14 verbose stack Error: dojo.io@1.0.0 test:ci: `run-s test:build test:build:static intern:ci test:clean`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (/Users/kanefreeman/.nvm/versions/node/v6.14.4/lib/node_modules/npm/lib/utils/lifecycle.js:255:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at ChildProcess.<anonymous> (/Users/kanefreeman/.nvm/versions/node/v6.14.4/lib/node_modules/npm/lib/utils/spawn.js:40:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:191:7)
+14 verbose stack     at maybeClose (internal/child_process.js:920:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:230:5)
+15 verbose pkgid dojo.io@1.0.0
+16 verbose cwd /Users/kanefreeman/Documents/GitHub/site
+17 error Darwin 17.7.0
+18 error argv "/Users/kanefreeman/.nvm/versions/node/v6.14.4/bin/node" "/Users/kanefreeman/.nvm/versions/node/v6.14.4/bin/npm" "run" "test:ci"
+19 error node v6.14.4
+20 error npm  v3.10.10
+21 error code ELIFECYCLE
+22 error dojo.io@1.0.0 test:ci: `run-s test:build test:build:static intern:ci test:clean`
+22 error Exit status 1
+23 error Failed at the dojo.io@1.0.0 test:ci script 'run-s test:build test:build:static intern:ci test:clean'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the dojo.io package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     run-s test:build test:build:static intern:ci test:clean
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs dojo.io
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls dojo.io
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
