+++ package.json
@@ -4,13 +4,15 @@
   "scripts": {
     "dev": "dojo build --mode dev --watch --serve",
     "intern": "intern",
+    "intern:ci": "intern config=@ci",
     "lint:tslint": "tslint -p .",
     "lint:prettier": "prettier -l \"{src,tests,scripts}/**/*.{ts,tsx}\"",
     "lint": "run-p lint:*",
     "prettier": "prettier --write \"{src,tests,scripts}/**/*.{ts,tsx}\"",
-    "test:build": "tsc",
+    "test:build": "tsc && dojo build -m unit",
     "test:build:static": "cpx \"{src,tests,content}/**/*.{html,png,jpg,gif,json,xml,css,txt,d.ts,svg,ico}\" _build",
     "test:clean": "shx rm -Rf _build",
+    "test:ci": "run-s test:build test:build:static intern:ci test:clean",
     "test": "run-s test:build test:build:static intern test:clean",
     "uploadCoverage": "codecov --file=output/coverage/coverage-final.json"
   },
