+++ scripts/tests/node.ts
@@ -1,0 +1,7 @@
+import './compile';
+import './logger';
+import './regions/parser';
+import './regions/region-matchers/block-c';
+import './regions/region-matchers/html';
+import './regions/region-matchers/inline-c';
+import './regions/region-matchers/inline-hash';
