+++ tests/unit/all.ts
@@ -1,16 +1,1 @@
-import './App';
-import './pages/Blog';
-import './pages/Community';
-import './pages/Documentation';
-import './pages/Examples';
-import './pages/Home';
-import './pages/Playground';
-import './widgets/Menu';
-
-import './scripts/compile';
-import './scripts/logger';
-import './scripts/regions/parser';
-import './scripts/regions/region-matchers/block-c';
-import './scripts/regions/region-matchers/html';
-import './scripts/regions/region-matchers/inline-c';
-import './scripts/regions/region-matchers/inline-hash';
+import './browser';
