+++ tests/unit/browser.ts
@@ -1,0 +1,8 @@
+import './App';
+import './pages/Blog';
+import './pages/Community';
+import './pages/Documentation';
+import './pages/Examples';
+import './pages/Home';
+import './pages/Playground';
+import './widgets/Menu';
