+++ tsconfig.json
@@ -28,6 +28,7 @@
 		"./src/**/*.ts",
 		"./src/**/*.tsx",
 		"./tests/**/*.ts",
-		"./tests/**/*.tsx"
+		"./tests/**/*.tsx",
+		"./scripts/**/*.ts"
 	]
 }
